Begin To Implement The Game World
Here I see a difference to freewar: I do not see a reason to maintain the grid layout from freewar. So I go with a simple graph to model the world. With this commit, the nodes part is added. To support movement, edges will be modeled too.
Also add view function to display the current state of the game world. Enable player to see where the character is located.