Improve Visuals For Mentoran Mountains
+ Use border to outline the shapes of individual peaks. This way, the same fill can be used over and over again, making this solution scale better.
+ Also add a visual hint for a planned future access to the mountain.
+ Record the map as exported from Gravit designer.

Some other things have changed in the map design due to a confusion of versions of the map design. Feels like I am spending a lot of time on the propagation of the map design from Gravit designer to the app. Better integration of Gravit with version control (to review changes before merging) would be nice.