Expand Map With Cave That Can Be Entered From Mentoran
+ Add a new group to map design to model the visuals and walking nodes in the cave.
+ Add text elements to describe where the player will be offered additional interactions, which allow him to enter and leave the cave.
+ Support localizing the source of unsupported path data in map SVG: Expand parsing function to log errors to console.
+ Expand parsing of map SVG to support negative coordinates for polygon points.

The plan is to use the text elements as a mechanism to model placement of structures in the game world. The text content will be parsed to determine what is placed. The idea for the two elements introduced with this changeset is to offer the particular interactions to the player when he is located on the walk node which is closest to the text element in the SVG.