Load in Gravit Designer and export as SVG
Using the editor at https://designer.gravit.io

I got this file by exporting it as SVG via the 'Advanced Export' dialog.

It looks like the new version of Gravit Designer broke the location representations. Despite the lock icons in the tree view in Gravit Designer, it changed the representation of the transform for the location nodes.